+++ WearAndTear/Config/Props/RegistryTemplates/ReinforcementTemplate.cs
@@ -10,7 +10,7 @@ namespace WearAndTear.Config.Props.RegistryTemplates
         /// This is used to fill the multiplier of the protective part of the Metal Reinforcement Template
         /// </summary>
         [DefaultValue(.95f)]
-        [Range(0, 1)]
+        [Range(0d, 1d)]
         [DisplayFormat(DataFormatString = "P")]
         public float DecayMultiplier { get; set; } = .95f;
 
@@ -18,7 +18,7 @@ namespace WearAndTear.Config.Props.RegistryTemplates
         /// This is used to fill the average life span of the Metal Reinforcement Template
         /// </summary>
         [DefaultValue(2)]
-        [Range(0, float.PositiveInfinity)]
+        [Range(0d, double.PositiveInfinity)]
         public float AvgLifeSpanInYears { get; set; } = 2;
     }
 }
